rename filenames to match js naming convention
